backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
